[fix] Compute _is_check_mode_active() by ourselves
Due to a prior refactoring with insufficient scrutinity, we were
calling the `_is_check_mode` method on the parent action to tell
whether check mode was active or not. But said method only exists in
some action classes of github.com/epfl-si/wp-ops.

- Fold the correct sensing code (based on the `ansible_check_mode`
task var) into a private method of Subaction