[fix][breaking?] Swap around Plan A and Plan B
It appears I was being confused as regards what `action_loader` resp. `.execute_module()` do. After wading through some code with the debugger, it appears as though the former runs an *action*  i.e. some kind of locally-loaded plugin to Ansible (whether author-contributed or from the Ansible core); while the latter uploads an AnsiballZ and runs it.

As it so happens that in the core, e.g. `ansible.builtin.copy` is both an action and a module; and only the former handles `content` correctly (i.e. creates a temp file and passes it around to the module); it follows that we should try to load an action plugin first.

AFAICT, there was no real reason for the priority order as it was written — I just remember that one had to try both. Hopefully this won't break any pre-existing caller code.